add basic chart

+++ app/src/main/java/com/codifilia/gotas/precipitation/Model.kt
@@ -4,9 +4,9 @@ import java.util.*
 
 data class Observation(val time: Date, val value: Precip?)
 
-sealed class Precip(open val amount: Double) {
-    class Rain(override val amount: Double) : Precip(amount)
-    class Snow(override val amount: Double) : Precip(amount)
-    class Mixed(override val amount: Double) : Precip(amount)
-    object None : Precip(0.0)
+sealed class Precip(open val amount: Float) {
+    class Rain(override val amount: Float) : Precip(amount)
+    class Snow(override val amount: Float) : Precip(amount)
+    class Mixed(override val amount: Float) : Precip(amount)
+    object None : Precip(0.0f)
 }
